* array.c, class.c, eval.c, file.c, gc.c, node.h, numeric.c,   object.c, parse.y, range.c, re.c, string.c, struct.c, util.c,   ext/socket/socket.c: merge from 1.7: treat long values as long   instead of int.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/branches/ruby_1_6@3003 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ class.c
@@ -633,7 +633,7 @@ rb_singleton_class(obj)
 	SPECIAL_SINGLETON(Qnil, rb_cNilClass);
 	SPECIAL_SINGLETON(Qfalse, rb_cFalseClass);
 	SPECIAL_SINGLETON(Qtrue, rb_cTrueClass);
-	rb_bug("unknown immediate %d", obj);
+	rb_bug("unknown immediate %ld", (long)obj);
     }
 
     DEFER_INTS;
